feat: Add get_by_email_daily_todo method to Query class
This commit adds a new method called get_by_email_daily_todo to the Query class in the Graphql/query.py file. This method retrieves test data by email. The method takes an email_user parameter and returns a list of DailyTodoType objects.

The get_by_email_daily_todo method is implemented in the DailyTodoService class in the service/dailyTodo.py file. It calls the get_by_email method of the DailyTodoRepository class to retrieve the test data by email. The retrieved data is then converted into a list of DailyTodoType objects and returned.

Additionally, the commit updates the main.py file to change the port number from 8080 to 8081 in the uvicorn.run() function